+++ ui/components/form-element/RELEASENOTES.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 - Changed the alignment of address form element. This fixes the alignment issue in cases where help text is present.
+- Record form .slds-form__item width is now fluid when parent container is less than 280px
 
 ### Fixed
 - Fixed (added) back bem syntax for backwards compatibility